Refactor and new open thing
