+++ Classes and Structs/PictureBoxAndInts.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Windows.Forms;
+
+namespace TwitchChatCoroutines.ClassesAndStructs
+{
+    struct PictureBoxAndInts
+    {
+        public PictureBox pb;
+        public Tuple<int, int> ints;
+    }
+}
