+++ Resources/splitter.png
