+++ TwitchChatCoroutines.csproj
@@ -50,30 +50,42 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="askChannel.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="askChannel.Designer.cs">
-      <DependentUpon>askChannel.cs</DependentUpon>
-    </Compile>
-    <Compile Include="Coroutines.cs" />
-    <Compile Include="Form1.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="Form1.Designer.cs">
-      <DependentUpon>Form1.cs</DependentUpon>
-    </Compile>
-    <Compile Include="Password.cs" />
-    <Compile Include="Program.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Structs.cs">
+    <Compile Include="Classes and Structs\MessageControl.cs" />
+    <Compile Include="Classes and Structs\PictureBoxAndInts.cs" />
+    <Compile Include="Classes and Structs\TwitchLabel.cs">
       <SubType>Component</SubType>
     </Compile>
-    <EmbeddedResource Include="askChannel.resx">
+    <Compile Include="Forms\askChannel.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Forms\askChannel.Designer.cs">
+      <DependentUpon>askChannel.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Classes and Structs\Coroutines.cs" />
+    <Compile Include="Forms\ChatForm.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Forms\ChatForm.Designer.cs">
+      <DependentUpon>ChatForm.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Classes and Structs\Password.cs" />
+    <Compile Include="Classes and Structs\Program.cs" />
+    <Compile Include="Forms\MainForm.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Forms\MainForm.Designer.cs">
+      <DependentUpon>MainForm.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Classes and Structs\TwitchMessage.cs" />
+    <EmbeddedResource Include="Forms\askChannel.resx">
       <DependentUpon>askChannel.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Form1.resx">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <EmbeddedResource Include="Forms\ChatForm.resx">
+      <DependentUpon>ChatForm.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Forms\MainForm.resx">
+      <DependentUpon>MainForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
@@ -101,8 +113,8 @@
   </ItemGroup>
   <ItemGroup>
     <Content Include="FodyWeavers.xml" />
-    <None Include="splitter2.png" />
-    <None Include="splitter.png" />
+    <None Include="Resources\splitter2.png" />
+    <None Include="Resources\splitter.png" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
